pyproject.toml
[project]
name = "resgates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.26.0",
    "bizdays>=1.0.16",
    "datetime>=5.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pymssql>=2.3.7",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
]
